:sparkles: Finalize first version of GoogleAssistant
