+++ platforms/platform-googleassistant/src/GoogleAssistant.ts
@@ -1,4 +1,11 @@
-import { Jovo } from '@jovotech/framework';
+import {
+  App,
+  DbPluginStoredElementsConfig,
+  HandleRequest,
+  Jovo,
+  JovoPersistableData,
+  Platform,
+} from '@jovotech/framework';
 import { GoogleAssistantResponse } from '@jovotech/output-googleassistant';
 import { GoogleAssistantDevice } from './GoogleAssistantDevice';
 import { GoogleAssistantPlatform } from './GoogleAssistantPlatform';
@@ -12,4 +19,28 @@ export class GoogleAssistant extends Jovo<
   GoogleAssistantUser,
   GoogleAssistantDevice,
   GoogleAssistantPlatform
-> {}
+> {
+  constructor($app: App, $handleRequest: HandleRequest, $platform: GoogleAssistantPlatform) {
+    super($app, $handleRequest, $platform);
+    if (this.$request.session?.params?._GA_REPROMPTS_) {
+      this.$session._GA_REPROMPTS_ = this.$request.session.params._GA_REPROMPTS_;
+    }
+  }
+
+  getPersistableData(): JovoPersistableData {
+    const persistableData = super.getPersistableData();
+    if (persistableData.session) {
+      persistableData.session._GA_REPROMPTS_ = this.$session._GA_REPROMPTS_;
+    }
+    return persistableData;
+  }
+
+  setPersistableData(data: JovoPersistableData, config?: DbPluginStoredElementsConfig) {
+    super.setPersistableData(data, config);
+    if ((typeof config?.session === 'object' && config.session.enabled) || config?.session) {
+      if (data.session?._GA_REPROMPTS_) {
+        this.$session._GA_REPROMPTS_ = data.session?._GA_REPROMPTS_;
+      }
+    }
+  }
+}
